requirements.txt
flask==1.1.1
requests==2.23.0
pyyaml==5.3.1
celery==4.4.2
fp-NGFW-SMC-python==0.7.0b20